Implement concurrent test for Vert.x http server (#2749)
* Implement concurrent test for Vert.x http server

* Format

+++ instrumentation/servlet/servlet-3.0/javaagent/src/test/groovy/TestServlet3.groovy
@@ -40,7 +40,7 @@ class TestServlet3 {
             break
           case INDEXED_CHILD:
             resp.status = endpoint.status
-            Span.current().setAttribute("test.request.id", req.getParameter("id") as int)
+            Span.current().setAttribute("test.request.id", req.getParameter("id") as long)
             break
           case QUERY_PARAM:
             resp.status = endpoint.status
